fix: UITextFields being inverted in landscape and reduced duplicated code
UITextFields were not displaying right when the
device is shifted to landscape orientation due to
me manually adjusting the constraints. To fix
I listened to the notification for when the device
orientation shifts and have it call the same code
that manually adjust the constraints.

I also deleted some duplicated code in
viewDidLoad and imagePickerController
to improve readability.